(1248) Split up planModel
