+++ packages/api/server/models/planModel/delete.js
@@ -1,0 +1,10 @@
+const { sequelize } = require('#db/models');
+
+module.exports = planId => sequelize.query(
+    'DELETE FROM plans WHERE plan_id = :planId',
+    {
+        replacements: {
+            planId,
+        },
+    },
+);
