+++ packages/api/server/models/planModel/findAll.js
@@ -1,0 +1,3 @@
+const query = require('./_common/query');
+
+module.exports = user => query(user, 'list');
